Fix spawn loader

+++ src/GameWorldSimulator/NeoServer.Game/Common/Contracts/World/ISpawn.cs
@@ -18,6 +18,6 @@ public interface ISpawn
 public interface ISpawnPoint
 {
     Location.Structs.Location Location { get; }
-    ushort SpawnTime { get; }
+    uint SpawnTime { get; }
     Direction Direction { get; set; }
 }
